refactor: Hospital Response 패키지 구분(#179)
