+++ data/src/main/java/com/example/data/hospital/remote/response/list/SimpleHospitalListResponse.kt
@@ -1,8 +1,9 @@
-package com.example.data.hospital.remote.response
+package com.example.data.hospital.remote.response.list
 
 import com.example.data.common.mapper.DataMapper
 import com.example.data.common.network.BaseResponse
-import com.example.data.hospital.remote.response.SimpleHospitalResponse.Companion.toDomainModel
+import com.example.data.hospital.remote.response.single.SimpleHospitalResponse
+import com.example.data.hospital.remote.response.single.SimpleHospitalResponse.Companion.toDomainModel
 import com.example.domain.hospital.model.SimpleHospitalList
 import com.google.gson.annotations.SerializedName
 import kotlinx.parcelize.Parcelize
